users/requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
flask_cors
bcrypt
uuid
python-dotenv